requirements.txt
absl-py==0.8.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.4
blingfire==0.0.11
boto==2.49.0
boto3==1.9.228
botocore==1.12.228
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.1.1
contextlib2==0.5.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dm-sonnet==1.32
docutils==0.15.2
entrypoints==0.3
future==0.17.1
gast==0.2.2
gensim==3.8.0
GetDist==1.0.3
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.23.0
h5py==2.10.0
html2text==2019.8.11
idna==2.8
imageio==2.6.1
importlib-metadata==1.6.0
ipdb==0.12
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.1.4
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.13.0
librosa==0.7.0
llvmlite==0.29.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
MulticoreTSNE==0.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.45.1
numexpr==2.7.0
numpy==1.17.2
packaging==19.2
pandas==0.25.0
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
pluggy==0.13.0
ply==3.11
progressbar2==3.43.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==3.0.5
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.6.1
pyGTC==0.3.1
pyparsing==2.4.0
pyrsistent==0.16.0
pytest==5.2.2
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.22.0
requests-file==1.4.3
resampy==0.2.2
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.0
seaborn==0.9.0
semantic-version==2.6.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==1.8.4
SoundFile==0.10.2
soupsieve==1.9.3
spherecluster==0.1.7
statsmodels==0.10.2
style==1.1.0
tables==3.5.2
tensorboard==1.14.0
tensorflow-datasets==1.2.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-metadata==0.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tldextract==2.2.1
torchfile==0.1.0
tornado==6.0.4
tqdm==4.35.0
traitlets==4.3.3
unicode==2.7
Unidecode==1.1.1
update==0.0.1
urllib3==1.25.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
